Implemented what I call a Shell Filter.
Allows none, some or all of a documents text to be run through the
standard input of any shell command, including pipes and redirection.
new document can also be created with the output. Or a new document
can be created with the output.  Should be very useful to anyone
familiar with text processing commands.  More so than just the embedded
terminal.