Fixed searchui's ReplaceAll function.
An obvious error I should never have made.  Each time a replacement
was made it changed the environment. Old ReplaceAll did not
take into account that all the offsets would change when the text
changed.  Fixed by simply replacing everything in reverse.
(This fix will have to be fixed if used in a right to left text
language I imagine)