Big debugger changes
Most visible is the disassembly window, its all I hoped it would be...
actually I wasn't going to do one, who really knows assembly these days?

Did it because it was simple (hahaha) Revealed plenty of errors going on
behind the scenes (errors and ignorance)
Hoping the remainder of what I want to do in the debugger goes a little
smoother.