Signals for document's focus and keypress events
Use phobos signals for focus out and keypresses.
Originally I intended to make switching from gtk easy so I was trying
to limit where gtk was used.  This is the legacy of that. I really need
to create a robust API for elements.  Also... pango does not recognize
an underscore as a regular word character, started working around that.