Improvements to ddoc to pango markup.
Certainly not professional quality but parameters are in a table and
code has a black background with a green font.
Gave up making tables look like tables.  Problem is getting line lengths
with macros and pango tags. I'll come back to it at some point.
Oh... and parameters are not in the correct order!