Trying to get scrolling right on the main pane.
Using trial and error (Terrible terrible)
Removed an unneeded scrolled window.  Too many levels of scrolling.
Experimenting on resize and shrink for the pane.
Goal is to make the notebook page any size but still be able to
scroll to any part.

Also documents started with an insane width! And typing past the
visible window lost the cursor instead of scrolling.